feat: align user controls with header
