+++ src/views/dashboard/Title.vue
@@ -1,10 +1,11 @@
 <script setup>
-import { useTheme } from 'vuetify';
-import { ref, onMounted, onUnmounted } from 'vue';
+import { ref, onMounted, onUnmounted, computed } from 'vue';
+import { useRouter } from 'vue-router';
 import NavbarThemeSwitcher from '@/layouts/components/NavbarThemeSwitcher.vue';
-import { VCol } from 'vuetify/components';
 
-const { global } = useTheme();
+const user = ref(null);
+const router = useRouter();
+const API_BASE = (import.meta.env.VITE_API_BASE_URL || 'https://ossprey.ngrok.app').replace(/\/$/, '');
 
 // Reactive variable to track viewport width
 const isMobileView = ref(window.innerWidth < 768);
@@ -17,11 +18,34 @@ const updateViewport = () => {
 // Add and remove event listener on component lifecycle
 onMounted(() => {
   window.addEventListener('resize', updateViewport);
+  const stored = localStorage.getItem('user');
+  if (stored) {
+    try {
+      user.value = JSON.parse(stored);
+    }
+    catch {
+      // ignore parse errors
+    }
+  }
 });
 
 onUnmounted(() => {
   window.removeEventListener('resize', updateViewport);
 });
+
+const userName = computed(() => user.value?.name || user.value?.email || '');
+
+const logout = async () => {
+  try {
+    await fetch(`${API_BASE}/api/logout`, { method: 'POST' });
+  }
+  catch {
+    // ignore errors
+  }
+  localStorage.removeItem('user');
+  user.value = null;
+  router.push('/login');
+};
 </script>
 
 <template>
@@ -41,10 +65,26 @@ onUnmounted(() => {
           <span v-else>OSSPREY (Open Source Software PRojEct sustainabilitY tracker)</span>
         </VCardTitle>
       </div>
-      
-      <!-- Theme Switcher -->
-      <div class="theme-switcher-container">
+
+      <!-- Action Icons -->
+      <div class="actions-container">
         <NavbarThemeSwitcher />
+
+        <VTooltip v-if="user" :text="userName" location="bottom">
+          <template #activator="{ props }">
+            <VBtn v-bind="props" icon class="ms-2">
+              <VIcon icon="bx-user" />
+            </VBtn>
+          </template>
+        </VTooltip>
+
+        <VTooltip v-if="user" text="Log Out" location="bottom">
+          <template #activator="{ props }">
+            <VBtn v-bind="props" icon class="ms-2" @click="logout">
+              <VIcon icon="bx-log-out" />
+            </VBtn>
+          </template>
+        </VTooltip>
       </div>
     </VCardText>
   </VCard>
@@ -62,16 +102,18 @@ onUnmounted(() => {
   text-align: center;
 }
 
-.theme-switcher-container {
+.actions-container {
   position: absolute;
-  right: 16px; 
+  right: 16px;
   top: 50%;
   transform: translateY(-50%);
+  display: flex;
+  align-items: center;
 }
 
 @media (max-width: 767px) {
   .title-container {
-    padding-right: 48px;
+    padding-right: 120px;
   }
 }
 </style>
